refactor: PieceDao 삭제, DataConfig로 상수 분리
