+++ src/main/java/janggi/Application.java
@@ -1,6 +1,5 @@
 package janggi;
 
-import janggi.data.PieceDao;
 import janggi.data.BoardDao;
 import janggi.view.InitializeView;
 import janggi.view.JanggiBoardView;
@@ -9,8 +8,7 @@ public class Application {
     public static void main(String[] args) {
         InitializeView initializeView = new InitializeView();
         JanggiBoardView janggiBoardView = new JanggiBoardView();
-        PieceDao pieceDao = new PieceDao();
-        BoardDao boardDao = new BoardDao(pieceDao);
+        BoardDao boardDao = new BoardDao();
         JanggiGame janggiGame = new JanggiGame(initializeView, janggiBoardView, boardDao);
         janggiGame.start();
     }
