+++ src/main/java/janggi/data/BoardDao.java
@@ -1,5 +1,11 @@
 package janggi.data;
 
+import static janggi.data.DataConfig.DATABASE;
+import static janggi.data.DataConfig.OPTION;
+import static janggi.data.DataConfig.PASSWORD;
+import static janggi.data.DataConfig.SERVER;
+import static janggi.data.DataConfig.USERNAME;
+
 import janggi.data.dto.BoardDto;
 import janggi.data.dto.PiecePointDto;
 import janggi.domain.board.Dynasty;
@@ -17,12 +23,6 @@ import java.util.Map;
 import java.util.Map.Entry;
 
 public final class BoardDao {
-
-    private static final String SERVER = "localhost:13306"; // MySQL 서버 주소
-    private static final String DATABASE = "chess"; // MySQL DATABASE 이름
-    private static final String OPTION = "?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC";
-    private static final String USERNAME = "root"; //  MySQL 서버 아이디
-    private static final String PASSWORD = "root"; // MySQL 서버 비밀번호
 
     private static final List<BoardDto> INITIAL_BOARD_DTOS = new ArrayList<>() {{
         add(new BoardDto(1, 2, 5));
@@ -51,12 +51,6 @@ public final class BoardDao {
         add(new BoardDto(10, 7, 9));
     }};
 
-    private final PieceDao pieceDao;
-
-    public BoardDao(PieceDao pieceDao) {
-        this.pieceDao = pieceDao;
-    }
-
     public void initializePiecePoints() {
         withConnection(connection -> {
             String query = "INSERT INTO board (piece_id, x, y) VALUES (?, ?, ?)";
@@ -74,7 +68,7 @@ public final class BoardDao {
 
     public void updatePiecePoints(Map<Point, Piece> boardPieces) {
         withConnection(connection -> {
-            Map<Piece, Integer> pieceKeys = pieceDao.getPieceKeys();
+            Map<Piece, Integer> pieceKeys = getPieceKeys();
             String query = "INSERT board SET x = ?, y = ?, piece_id = ?";
             try (final var preparedStatement = connection.prepareStatement(query)) {
                 for (Entry<Point, Piece> entry : boardPieces.entrySet()) {
@@ -129,6 +123,24 @@ public final class BoardDao {
                 preparedStatement.executeUpdate();
             }
         });
+    }
+
+    public Map<Piece, Integer> getPieceKeys () {
+        Map<Piece, Integer> pieceKeys = new HashMap<>();
+        String query = "SELECT piece_id, piece_type, dynasty FROM piece";
+        try (final var connection = getConnection();
+             final var preparedStatement = connection.prepareStatement(query);
+             ResultSet resultSet = preparedStatement.executeQuery()) {
+            while(resultSet.next()) {
+                PieceType type = PieceType.valueOf(resultSet.getString("piece_type"));
+                Dynasty dynasty = Dynasty.valueOf(resultSet.getString("dynasty"));
+                int key = resultSet.getInt("piece_id");
+                pieceKeys.put(type.from(dynasty), key);
+            }
+        } catch (final SQLException e) {
+            throw new RuntimeException(e);
+        }
+        return pieceKeys;
     }
 
     private void withConnection(SqlConsumer<Connection> consumer) {
