+++ src/main/java/janggi/data/DataConfig.java
@@ -1,0 +1,9 @@
+package janggi.data;
+
+public class DataConfig {
+    protected static final String SERVER = "localhost:13306"; // MySQL 서버 주소
+    protected static final String DATABASE = "chess"; // MySQL DATABASE 이름
+    protected static final String OPTION = "?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC";
+    protected static final String USERNAME = "root"; //  MySQL 서버 아이디
+    protected static final String PASSWORD = "root"; // MySQL 서버 비밀번호
+}
